+++ src/main/java/finalmission/customer/controller/TokenLoginController.java
@@ -1,7 +1,7 @@
 package finalmission.customer.controller;
 
-import finalmission.customer.controller.dto.TokenLoginCreateRequest;
-import finalmission.customer.controller.dto.TokenLoginResponse;
+import finalmission.customer.controller.dto.request.TokenLoginCreateRequest;
+import finalmission.customer.controller.dto.response.TokenLoginResponse;
 import finalmission.customer.entity.Customer;
 import finalmission.customer.resolver.LoginCustomer;
 import finalmission.customer.service.AuthService;
