+++ src/main/java/finalmission/customer/controller/dto/response/TokenLoginResponse.java
@@ -1,4 +1,4 @@
-package finalmission.customer.controller.dto;
+package finalmission.customer.controller.dto.response;
 
 public record TokenLoginResponse(
         String token
