+++ src/main/java/finalmission/reservatioin/service/ReservationService.java
@@ -4,8 +4,8 @@ import finalmission.customer.entity.Customer;
 import finalmission.customer.repository.CustomerJpaRepository;
 import finalmission.omakase.entity.Omakase;
 import finalmission.omakase.repository.OmakaseJpaRepository;
-import finalmission.reservatioin.controller.dto.CurrentStateReservationResponse;
-import finalmission.reservatioin.controller.dto.ReservationCreateRequest;
+import finalmission.reservatioin.controller.dto.response.CurrentStateReservationResponse;
+import finalmission.reservatioin.controller.dto.request.ReservationCreateRequest;
 import finalmission.reservatioin.entity.Reservation;
 import finalmission.reservatioin.entity.ReservationTime;
 import finalmission.reservatioin.entity.ReservationWithNumberOfPeople;
